+++ src/frontend/src/models/state/TestRunState.ts
@@ -1,8 +1,10 @@
 import ResultGroups from "./ResultGroups";
+import SummaryState from "./SummaryState";
 
 interface TestRunState {
     filter: string;
     resultGroups: ResultGroups;
+    summary: SummaryState
 }
 
 export default TestRunState;