+++ src/frontend/src/models/trx/ResultSummary.ts
@@ -1,6 +1,0 @@
-
-interface ResultSummary {
-    outcome: string;
-}
-
-export default ResultSummary;