+++ src/frontend/src/models/trx/RunInfo.ts
@@ -1,0 +1,5 @@
+interface RunInfo {
+    timeStamp: Date
+}
+
+export default RunInfo;