+++ src/frontend/src/models/trx/Summary.ts
@@ -1,0 +1,14 @@
+import RunInfo from "./RunInfo";
+
+interface Summary {
+    outcome: string;
+    counters: {
+        total: number;
+        executed: number;
+        passed: number;
+        failed: number;
+    },
+    runInfos: RunInfo[];
+}
+
+export default Summary;